renamed to bellman ford specific
